Seperate errors from routes Create errors.py and move error handlers to the file Remove unused imports from routes.py Add a config to store the token expiry time in config.py Changed expires_in to use time from config.py in helpers.py
